requirements.txt
pandas
scikit-learn
xgboost
joblib
PyGithub
openai
atlassian-python-api
streamlit
python-dotenv
pymongo
